+++ src/routes/components/InitialRoute.js
@@ -1,10 +1,10 @@
 import PropType from 'prop-types';
 import { Redirect, Route } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import Layout from '../components/common/Layout';
+import Layout from '../../components/common/Layout';
 
-const PrivateRoute = ({ component: Component, ...rest }) => {
-  const isAuthenticated = useSelector(state => state.auth.isAuth);
+const InitialRoute = ({ component: Component, ...rest }) => {
+  const isInitial = useSelector(state => state.initialUser.isInitial);
 
   return (
     <Layout>
@@ -13,15 +13,15 @@ const PrivateRoute = ({ component: Component, ...rest }) => {
         {...rest}
         render={props =>
           // eslint-disable-next-line
-          isAuthenticated ? <Component {...props} /> : <Redirect to="/" />
+          isInitial ? <Component {...props} /> : <Redirect to="/" />
         }
       />
     </Layout>
   );
 };
 
-export default PrivateRoute;
+export default InitialRoute;
 
-PrivateRoute.propTypes = {
+InitialRoute.propTypes = {
   component: PropType.func.isRequired
 };
