+++ src/routes/components/PrivateRoute.js
@@ -1,10 +1,10 @@
 import PropType from 'prop-types';
 import { Redirect, Route } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import Layout from '../components/common/Layout';
+import Layout from '../../components/common/Layout';
 
-const InitialRoute = ({ component: Component, ...rest }) => {
-  const isInitial = useSelector(state => state.initialUser.isInitial);
+const PrivateRoute = ({ component: Component, ...rest }) => {
+  const isAuthenticated = useSelector(state => state.auth.isAuth);
 
   return (
     <Layout>
@@ -13,15 +13,15 @@ const InitialRoute = ({ component: Component, ...rest }) => {
         {...rest}
         render={props =>
           // eslint-disable-next-line
-          isInitial ? <Component {...props} /> : <Redirect to="/" />
+          isAuthenticated ? <Component {...props} /> : <Redirect to="/" />
         }
       />
     </Layout>
   );
 };
 
-export default InitialRoute;
+export default PrivateRoute;
 
-InitialRoute.propTypes = {
+PrivateRoute.propTypes = {
   component: PropType.func.isRequired
 };
